+++ src/main/java/com/example/SVC/repository/DocumentRepository.java
@@ -2,6 +2,7 @@ package com.example.SVC.repository;
 
 import com.example.SVC.model.Document;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -13,6 +14,6 @@ public interface DocumentRepository extends JpaRepository<Document, Long> {
     Optional<Document> findByTitle(String title);
     
     @Query("SELECT MAX(dv.version) FROM DocumentVersion dv WHERE dv.document.title = :title")
-    Double findNewestVersion(@Param("title") String title);
+    BigDecimal findNewestVersion(@Param("title") String title);
     
 }
